file naming refactore and finalize Readme content
